requirements.txt
datasets>=3.0.0
tree-sitter>=0.23.0
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.23.0
tree-sitter-typescript>=0.23.0
tree-sitter-rust>=0.23.0
tree-sitter-go>=0.23.0
tree-sitter-java>=0.23.0
tqdm>=4.66.0
huggingface-hub>=0.25.0
pyyaml>=6.0.0
Pillow>=10.0.0
Pygments>=2.17.0
